// MegabinFE/src/app/services/model/customerScheduleContractDto.ts
/**
 * Megabin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DayOfWeek } from './dayOfWeek';
import { Frequency } from './frequency';


export interface CustomerScheduleContractDto { 
    id: string;
    addressId: string;
    address: string | null;
    frequency: Frequency;
    dayOfWeek: DayOfWeek;
    startingDate: string;
    active: boolean;
    approvedExternally: boolean;
}
export namespace CustomerScheduleContractDto {
}